api/src/main/java/top/jionjion/api/config/RedisConfig.java
package top.jionjion.api.config;

import com.fasterxml.jackson.annotation.JsonAutoDetect;
import com.fasterxml.jackson.annotation.PropertyAccessor;
import com.fasterxml.jackson.databind.ObjectMapper;
import lombok.AllArgsConstructor;
import lombok.Getter;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.cache.CacheManager;
import org.springframework.cache.annotation.CachingConfigurerSupport;
import org.springframework.cache.annotation.EnableCaching;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.data.redis.cache.RedisCacheManager;
import org.springframework.data.redis.connection.RedisConnection;
import org.springframework.data.redis.connection.RedisConnectionFactory;
import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
import org.springframework.data.redis.connection.jedis.JedisClientConfiguration;
import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
import org.springframework.data.redis.core.RedisTemplate;
import org.springframework.data.redis.core.StringRedisTemplate;
import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;

import javax.crypto.KeyGenerator;
import java.time.Duration;

/**
 * @author Jion
 *  Reidis数据库的配置类.
 *  需要自定义Bean完成配置
 */

@Configuration
@EnableCaching
public class RedisConfig extends CachingConfigurerSupport {

    @Value("${spring.redis.host}")
    private String host;

    @Value("${spring.redis.port}")
    private int port;

    @Value("${spring.redis.timeout}")
    private int timeout;

    @Value("${spring.redis.password}")
    private String password;

    @Value("${spring.redis.database}")
    private int database;



    /** 创建连接工厂 */
    @Bean
    public JedisConnectionFactory redisConnectionFactory() {
        // 连接配置类
        RedisStandaloneConfiguration redisStandaloneConfiguration = new RedisStandaloneConfiguration();
        redisStandaloneConfiguration.setHostName(host);
        redisStandaloneConfiguration.setPort(port);
        redisStandaloneConfiguration.setPassword(password);
        redisStandaloneConfiguration.setDatabase(database);

        // Jedis连接配置
        JedisClientConfiguration jedisClientConfiguration = JedisClientConfiguration.defaultConfiguration();
        JedisClientConfiguration.builder().connectTimeout(Duration.ofMillis(timeout));

        // 根据配置创建连接工厂
        return new JedisConnectionFactory(redisStandaloneConfiguration,jedisClientConfiguration);

    }

    /** 重新配置 RedisTemplate */
    public RedisTemplate redisTemplate(){
        JedisConnectionFactory jedisConnectionFactory = redisConnectionFactory();
        RedisTemplate redisTemplate = new RedisTemplate();
        redisTemplate.setConnectionFactory(jedisConnectionFactory);
        return redisTemplate;
    }
}